Add tests for missing count in tests for empty lists
In the following factories, there are missing tests for
the count of a list in the if-statements meant to ensure
that empty lists are not accessed. In three locations, the
result from the AQL query was not testes for `Count != 0`,
so that non-existent list elements were queried downstream,
leading to crashes.