Filter distinct the list of patient movements
The list of patient movements, in PatientMovementPs, that
is returned by EHRbase using the provided test data may in some
scenarios return wholly duplicate elements. This is rejected by the
data parser within SmICSVisualisierung. Since entirely duplicate
elements do not add information, they can be filtered in the backend,
before returning data to SmICSVisualisierung.

This implementation leverages LINQ's `Distinct()` operation, which
uses the provided equality comparator. To have a simple implementation
for `Equals` and `GetHashCode`, the serialization to JSON can be
employed - if the JSON of two instances of `PatientMovementModel` is
identical, the instances will be equal also.