Workaround for EHRbase AQL bug
The EHRbase currently has a bug in the AQL parser, where the sort order
(`ASC`, `DSC`, `ASCENDING` or `DESCENDING`) is non-optional:
https://github.com/ehrbase/ehrbase/issues/689

The AQL spec states that an ascending sort order is implied if not
otherwise specified, so this is clearly a bug in the EHRbase.

This bug leads to the rejection of the AQLQuery for the contact network.
By adding an explicit `ASC` sort to the respective AQL query, the issue
can be circumvented. Being more explicit should not impact compatibility
with the Better platform, so that the workaround should be permanently
in-place.